+++ packages/vite/src/node/http.ts
@@ -5,7 +5,7 @@ import type {
   OutgoingHttpHeaders as HttpServerHeaders
 } from 'node:http'
 import type { ServerOptions as HttpsServerOptions } from 'node:https'
-import type { Connect } from 'types/connect'
+import type { Connect } from 'dep-types/connect'
 import colors from 'picocolors'
 import { isObject } from './utils'
 import type { ProxyOptions } from './server/middlewares/proxy'
@@ -121,8 +121,7 @@ export async function resolveHttpServer(
 }
 
 export async function resolveHttpsConfig(
-  https: boolean | HttpsServerOptions | undefined,
-  cacheDir: string
+  https: boolean | HttpsServerOptions | undefined
 ): Promise<HttpsServerOptions | undefined> {
   if (!https) return undefined